@@ -4,12 +4,12 @@
     "type": "asgard-module",
     "license": "MIT",
     "require": {
-        "php": "^7.1.3",
-        "imagina/core-module": "v8.x-dev"
+        "php": "^8.1",
+        "imagina/core-module": "v10.x-dev"
     },
     "require-dev": {
-        "phpunit/phpunit": "~4.0",
-        "orchestra/testbench": "~3.2"
+        "orchestra/testbench": "^8.5",
+        "phpunit/phpunit": "^10"
     },
     "autoload-dev": {
         "psr-4": {
@@ -28,7 +28,7 @@
     "extra": {
         "branch-alias": {
             "dev-master": "4.0.x-dev",
-            "dev-v8.x": "8.0.x-dev"
+            "dev-v10.x": "10.0.x-dev"
         }
     },
     "autoload": {
