@@ -1,5 +1,8 @@
 <?php
 
+use App\Http\Controllers\IwhmcsController;
+use Illuminate\Support\Facades\Route;
+
 /*
 |--------------------------------------------------------------------------
 | Web Routes
@@ -11,6 +14,6 @@
 |
 */
 
-Route::prefix('iwhmcs')->group(function() {
-    Route::get('/', 'IwhmcsController@index');
+Route::prefix('iwhmcs')->group(function () {
+    Route::get('/', [IwhmcsController::class, 'index']);
 });
