@@ -2,28 +2,25 @@
 
 namespace Modules\Iwhmcs\Database\Seeders;
 
-use Illuminate\Database\Seeder;
 use Illuminate\Database\Eloquent\Model;
+use Illuminate\Database\Seeder;
 
 class IwhmcsModuleTableSeeder extends Seeder
 {
-  /**
-   * Run the database seeds.
-   *
-   * @return void
-   */
-  public function run()
-  {
-    Model::unguard();
-  
+    /**
+     * Run the database seeds.
+     */
+    public function run()
+    {
+        Model::unguard();
 
-    $columns = [
-      ["config" => "config", "name" => "config"],
-      ["config" => "permissions", "name" => "permissions"],
-    ];
-  
-    $moduleRegisterService = app("Modules\Isite\Services\RegisterModuleService");
-  
-    $moduleRegisterService->registerModule("iwhmcs", $columns, 1);
-  }
+        $columns = [
+            ['config' => 'config', 'name' => 'config'],
+            ['config' => 'permissions', 'name' => 'permissions'],
+        ];
+
+        $moduleRegisterService = app("Modules\Isite\Services\RegisterModuleService");
+
+        $moduleRegisterService->registerModule('iwhmcs', $columns, 1);
+    }
 }
